requirements.txt
asgiref==3.8.1
autoflake==2.3.1
autopep8==2.3.2
beautifulsoup4==4.12.3
black==25.1.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.8
cognitive_complexity==1.3.0
cssselect2==0.7.0
Django==4.2.17
django-check-seo==1.0.1
django-ckeditor==6.7.1
django-classy-tags==4.1.0
django-cms==4.1.4
django-filer==3.3.1
django-formtools==2.5.1
django-js-asset==2.2.0
django-meta==2.4.2
django-modeltranslation==0.17.6
django-polymorphic==3.1.0
django-rosetta==0.10.1
django-sekizai==4.1.0
django-treebeard==4.7.1
djangocms-admin-style==3.3.1
djangocms-page-meta==1.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
easy-thumbnails==2.9
flake8==7.1.2
Flake8-pyproject==1.2.3
flake8_cognitive_complexity==0.1.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
isort==6.0.1
lxml==5.3.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
polib==1.2.0
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
reportlab==4.2.5
requests==2.32.3
rest-framework-simplejwt==0.0.2
setuptools==75.8.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
svglib==1.5.1
tinycss2==1.4.0
transliterate==1.10.2
typing_extensions==4.12.2
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.3
webencodings==0.5.1
